+++ app/src/main/java/org/student/travelnoteapp/presentation/booking/BookingEditScreen.kt
@@ -8,7 +8,7 @@ import androidx.navigation.NavController
 fun BookingEditScreen(
     navController: NavController,
     viewModel: BookingDetailsViewModel = hiltViewModel(),
-    id: Int
+    id: Long
 ) {
 
 }
